publish prod branch to production server
If PUBLISH_TO_GITHUB is not defined, publish to our server with:
* git checkout prod; git merge master; make build
* git commit -a -m... git push; git checkout master

The recipe will fail if the working tree isn't clean or if jekyll serve
is running.  The latter would do a spurious build after the checkout.

Uploading with rsync is definitely cleaner, simpler, and easier on your
repository, but there may be situations where you need it.  If you want
to mix the two, you will need to switch back to master on the server.