Make a generic collection layout
Move most of what was in Articles.html into _layouts/collection.html,
which now takes page content.  Specify the collection, and list title if
any, in the frontmatter.

The main difference between the collection and blog layout is that the
blog is sorted in reverse, to put the most recent posts at the top.